requirements.txt
absl-py==2.1.0
accelerate==1.1.1
advertorch==0.2.4
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
brisque==0.0.16
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
clarabel==0.9.0
click==8.1.7
clip==1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cvxopt==1.3.2
cvxpy==1.6.0
cycler==0.12.1
cython==3.0.11
debugpy==1.8.8
decorator==5.1.1
diffusers==0.25.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==5.5.0
gradio_client==1.4.2
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
insightface==0.7.3
intel-cmplr-lib-ur==2025.0.2
intel-openmp==2025.0.2
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.10.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
libclang==18.1.1
libsvm-official==3.35.0
lightning-utilities==0.11.8
lz4==4.3.3
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2025.0.1
ml_collections==1.0.0
mosek==10.2.8
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
peft==0.13.2
pexpect==4.9.0
photomaker==0.2.0
piexif==1.1.3
pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytorch-fid==0.3.0
pytorch-lightning==1.9.0
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scs==3.2.7
semantic-version==2.10.0
sentry-sdk==2.18.0
setproctitle==1.3.3
shellingham==1.5.4
simsimd==6.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spaces==0.30.4
stack-data==0.6.3
starlette==0.41.2
stringzilla==3.10.9
sympy==1.13.1
taming-transformers==0.0.1
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.12
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.25.1
triton==3.1.0
typer==0.13.0
typing_extensions==4.12.2
umf==0.9.1
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.1
youtube-dl==2021.12.17
zipp==3.20.2

